Update to Expand 'gene_match' Function
This now includes using the (long-form) 'gene_name' field, while allowing for extra characters at before, in-between, and at the end of the gene name to consider it a 'match'. I did this because I found records where there were '-' at the end, and the existing defined parsing pattern for 'gene_symbol' purposely did not allow for that. Due to the likelihood of false positives for the 'gene_name', the parsing pattern was uniquely modified for it.