Changed Web Scrapping to Monthly Occurrence
After speaking with Dr. Al-Chalabi and receiving communication regarding the frequency of updates to the ALSoD website, I have modified the frequency in which the web scrapping occurs. The function now checks the newly created Update_Log database table and, if the Gene table's last updated date was within 30 days of the current date, the function retrieves the Gene database's existing data instead of scrapping for an update.